LayoutTests/fast/dom/exception-in-binding-expected.txt

Tests to ensure that attributes are not set when an exceptions is raised while converting argument to JSValue.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS select.name is "select"
PASS select.selectedIndex is 1
